PythonTradingBot/requirements.txt
Python-binance
TA-Lib
numpy
websocket_client